fitlearn-platform/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.3.3
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2.0
pytest==7.2.2
pytest-django==4.5.2
gunicorn==21.2.0
django-environ==0.8.1